+++ templates/base/@core/context/SettingsContext.tsx
@@ -1,0 +1,103 @@
+'use client'
+
+// ** React Imports
+import { createContext, useState, ReactNode, useEffect } from 'react'
+
+// ** MUI Imports
+import { Direction } from '@mui/material'
+
+// ** ThemeConfig Import
+import themeConfig from '../configs/themeConfig'
+
+// ** Types Import
+import { Mode, ThemeColor } from '../layouts/types'
+
+export type Settings = {
+  mode: Mode
+  direction: Direction
+  themeColor: ThemeColor
+  toastPosition?: 'top-left' | 'top-center' | 'top-right' | 'bottom-left' | 'bottom-center' | 'bottom-right'
+}
+
+export type PageSpecificSettings = {
+  mode?: Mode
+  direction?: Direction
+  themeColor?: ThemeColor
+  toastPosition?: 'top-left' | 'top-center' | 'top-right' | 'bottom-left' | 'bottom-center' | 'bottom-right'
+}
+export type SettingsContextValue = {
+  settings: Settings
+  saveSettings: (updatedSettings: Settings) => void
+}
+
+interface SettingsProviderProps {
+  children: ReactNode
+  pageSettings?: PageSpecificSettings | void
+}
+
+const initialSettings: Settings = {
+  themeColor: 'primary',
+  mode: themeConfig.mode,
+  direction: themeConfig.direction,
+  toastPosition: themeConfig.toastPosition
+}
+
+const staticSettings = {
+  toastPosition: initialSettings.toastPosition
+}
+
+const restoreSettings = (): Settings | null => {
+  let settings = null
+
+  try {
+    const storedData: string | null = window.localStorage.getItem('settings')
+
+    if (storedData) {
+      settings = { ...JSON.parse(storedData), ...staticSettings }
+    } else {
+      settings = initialSettings
+    }
+  } catch (err) {
+    console.error(err)
+  }
+
+  return settings
+}
+
+// set settings in localStorage
+const storeSettings = (settings: Settings) => {
+  const initSettings = Object.assign({}, settings)
+
+  delete initSettings.toastPosition
+  window.localStorage.setItem('settings', JSON.stringify(initSettings))
+}
+
+// ** Create Context
+export const SettingsContext = createContext<SettingsContextValue>({
+  saveSettings: () => null,
+  settings: initialSettings
+})
+
+export const SettingsProvider = ({ children, pageSettings }: SettingsProviderProps) => {
+  const [settings, setSettings] = useState<Settings>({ ...initialSettings })
+
+  useEffect(() => {
+    const restored = restoreSettings()
+    const next = restored || initialSettings
+
+    if (pageSettings) {
+      setSettings({ ...next, ...pageSettings })
+    }
+
+    setSettings(next)
+  }, [pageSettings])
+
+  const saveSettings = (updatedSettings: Settings) => {
+    storeSettings(updatedSettings)
+    setSettings(updatedSettings)
+  }
+
+  return <SettingsContext.Provider value={{ settings, saveSettings }}>{children}</SettingsContext.Provider>
+}
+
+export const SettingsConsumer = SettingsContext.Consumer